feat: moved some common code from abfe.py to fep.py
